tests/wpt/mozilla/tests/mozilla/bluetooth/getCharacteristics/get-same-characteristics.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => Promise.all([
        service.getCharacteristics(body_sensor_location.name),
        service.getCharacteristics(body_sensor_location.name)]))
    .then(characteristics => {
        let chars1 = characteristics[0];
        let chars2 = characteristics[1];
        assert_equals(chars1.length, chars2.length);
        for (let i = 0; i < chars1.length; i++) {
            assert_equals(chars1[i], chars2[i]);
        }
    });
}, 'Calls to get the same characteristics should return the same objects.');
</script>
